Version 0.2 [Rough]
This version has basic functionality but it's not fully there yet. There are bugs, including weird crashing when the program ends (likely caused by calling `delete[]` on the same memory address twice. Also, the simulation logic or implementation is bad, causing the game of life rules to not apply correctly (this can be tested by creating a 3 x 1 or 1 x 3 configuration.